[tracing] ftrace: Add check for NULL regs if ops has SAVE_REGS set
Message-id: <1381413926-15908-492-git-send-email-vfalico@redhat.com>
Patchwork-id: 69063
O-Subject: [RHEL7 PATCH 491/820] ftrace: Add check for NULL regs if ops has SAVE_REGS set
Bugzilla: 1006938
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ: 1006938
Upstream: 195a8afc7ac962f8da795549fe38e825f1372b0d
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=195a8afc7ac962f8da795549fe38e825f1372b0d
Stable: 29632b10ef02a31dec2918177f6d4244e66d635d
http://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/?id=29632b10ef02a31dec2918177f6d4244e66d635d

Description
----------------------------
Backport from linux-3.10.y stable.

From: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Subject: ftrace: Add check for NULL regs if ops has SAVE_REGS set

commit 195a8afc7ac962f8da795549fe38e825f1372b0d upstream.

If a ftrace ops is registered with the SAVE_REGS flag set, and there's
already a ops registered to one of its functions but without the
SAVE_REGS flag, there's a small race window where the SAVE_REGS ops gets
added to the list of callbacks to call for that function before the
callback trampoline gets set to save the regs.

The problem is, the function is not currently saving regs, which opens
a small race window where the ops that is expecting regs to be passed
to it, wont. This can cause a crash if the callback were to reference
the regs, as the SAVE_REGS guarantees that regs will be set.

To fix this, we add a check in the loop case where it checks if the ops
has the SAVE_REGS flag set, and if so, it will ignore it if regs is
not set.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit 29632b10ef02a31dec2918177f6d4244e66d635d)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>