[tracing] uprobes: Fail to unregister if probe event files are in use
Message-id: <1381413926-15908-501-git-send-email-vfalico@redhat.com>
Patchwork-id: 69155
O-Subject: [RHEL7 PATCH 500/820] tracing/uprobes: Fail to unregister if probe event files are in use
Bugzilla: 1006938
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ: 1006938
Upstream: c6c2401d8bbaf9edc189b4c35a8cb2780b8b988e
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=c6c2401d8bbaf9edc189b4c35a8cb2780b8b988e
Stable: 396dd500521287215d836b84b2ae61cd9163bba5
http://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/commit/?id=396dd500521287215d836b84b2ae61cd9163bba5

Description
----------------------------
Backport from linux-3.10.y stable.

From: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Subject: tracing/uprobes: Fail to unregister if probe event files are in use

commit c6c2401d8bbaf9edc189b4c35a8cb2780b8b988e upstream.

Uprobes suffer the same problem that kprobes have. There's a race between
writing to the "enable" file and removing the probe. The probe checks for
it being in use and if it is not, goes about deleting the probe and the
event that represents it. But the problem with that is, after it checks
if it is in use it can be enabled, and the deletion of the event (access
to the probe) will fail, as it is in use. But the uprobe will still be
deleted. This is a problem as the event can reference the uprobe that
was deleted.

The fix is to remove the event first, and check to make sure the event
removal succeeds. Then it is safe to remove the probe.

When the event exists, either ftrace or perf can enable the probe and
prevent the event from being removed.

Link: http://lkml.kernel.org/r/20130704034038.991525256@goodmis.org

Acked-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit 396dd500521287215d836b84b2ae61cd9163bba5)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>