[security] keys: Make BIG_KEYS boolean
Message-id: <20131018173948.25139.52216.stgit@warthog.procyon.org.uk>
Patchwork-id: 69666
O-Subject: [RHEL7 PATCH 1/3] KEYS: Make BIG_KEYS boolean [BZ 991110]
Bugzilla: 991110
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Josh Boyer <jwboyer@redhat.com>

Having the big_keys functionality as a module is very marginally useful.
The userspace code that would use this functionality will get odd error
messages from the keys layer if the module isn't loaded.  The code itself
is fairly small, so just have this as a boolean option and not a tristate.

This is a fix for the patches submitted under BZ 991110.

Signed-off-by: Josh Boyer <jwboyer@fedoraproject.org>
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>