[netdrv] cnic: Fix crash in cnic_bnx2x_service_kcq()
Message-id: <1382366342-11837-2-git-send-email-mlombard@redhat.com>
Patchwork-id: 69747
O-Subject: [RHEL7 e-stor PATCH 1/1] cnic: Fix crash in cnic_bnx2x_service_kcq()
Bugzilla: 1011203
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

commit 104a43edb264321a4d41850e98153b4fa8a9ef42
    cnic: Use CHIP_NUM macros from bnx2x.h

changed the code to use the bnx2x macro NO_FCOE() to determine if FCoE
is supported or not.  There is another place in cnic that is still using
the old method to determine if FCoE is supported or not.  The 2 methods
may not yield the same result after the network interface is brought down
and up.  This will cause the crash as cnic_bnx2x_service_kcq() will access
the uninitialized cp->kcq2.

The fix is to consistently use the same macro CNIC_SUPPORTS_FCOE() which
uses the bnx2x NO_FCOE() macro.  As a follow-up, we can clean up the code
to remove the old method as it is no longer needed.

Signed-off-by: Michael Chan <mchan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 48a30569732be782119960a47ca05417e7e07dab)

Signed-off-by: Jarod Wilson <jarod@redhat.com>