[ethernet] ixgbevf: move API neg to reset path
Message-id: <1383249597-29425-1-git-send-email-gospo@redhat.com>
Patchwork-id: 70077
O-Subject: [PATCH RHEL7] ixgbevf: move API neg to reset path
Bugzilla: 1023107
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

It was discovered during RHEL6 testing that ixgbevf-based devices that
were assigned to guests could not set their MTU until link was up.  This
had to do with a way the PF<->VF communication channel was setup and
some poor assumptions made during the first attempt to negotiate this
properly.  Since the ixgbevf drivers were of similar vintage in RHEL6.5
and RHEL7 it is also a problem on RHEL7 and needs to be fixed.

A backport of the following upstream commit resolves this issue:

	commit 798e381a0c387dcbcb014e07c0e6683cf2d2dc22
	Author: Don Skidmore <donald.c.skidmore@intel.com>
	Date:   Tue Oct 1 04:33:51 2013 -0700

	    ixgbevf: move API neg to reset path

Fix confirmed by me to resolve RHBZ 1023107.

For those interested, the brew build is available here:
https://brewweb.devel.redhat.com/taskinfo?taskID=6463242

Signed-off-by: Jarod Wilson <jarod@redhat.com>