[virt] kvm: Add VFIO device
Message-id: <20131101183325.22449.68667.stgit@bling.home>
Patchwork-id: 70135
O-Subject: [RHEL7.0 kernel PATCH 1/3] kvm: Add VFIO device
Bugzilla: 1025470
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Bugzilla: 1025470
Upstream commit: ec53500fae421e07c5d035918ca454a429732ef4 (kvm.git/next)

So far we've succeeded at making KVM and VFIO mostly unaware of each
other, but areas are cropping up where a connection beyond eventfds
and irqfds needs to be made.  This patch introduces a KVM-VFIO device
that is meant to be a gateway for such interaction.  The user creates
the device and can add and remove VFIO groups to it via file
descriptors.  When a group is added, KVM verifies the group is valid
and gets a reference to it via the VFIO external user interface.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>