[drm] qxl: avoid an oops in the deferred io code
Message-id: <1383547895-1225-8-git-send-email-airlied@redhat.com>
Patchwork-id: 70482
O-Subject: [PATCH 7/7] [PATCH RHEL7.0] qxl: avoid an oops in the deferred io code.
Bugzilla: 1003728 1026182
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>

bugzilla: 1003728
upstream: send to dri-devel mailing list.

If we are using deferred io due to plymouth or X.org fbdev driver
we will oops in memcpy due to this pointless multiply here,

removing it fixes fbdev to start and not oops.

Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>