[net] nf_tables: stuff structures to preserve kABI in the future
Message-id: <b17673ebb2d126094da663d8c241cc36393cd964.1383605196.git.jbenc@redhat.com>
Patchwork-id: 70527
O-Subject: [RHEL7 net PATCH 21/23] nf_tables: stuff structures to preserve kABI in the future
Bugzilla: 1023123
RH-Acked-by: Jesper Brouer <jbrouer@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Daniel Borkmann <dborkman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1023123
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6523956
Upstream: RHEL specific

Struct net is on the kABI whitelist through struct net_device and its
numerous users.

Nothing using nf_hook_ops is on the whitelist currently but as it was on the
whitelist in RHEL6, it's likely it will be extended and it's unlikely we'll
allow it to the whitelist in the near future, let's be nice to its users.

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>