[virt] hyperv/vmbus: Implement multi-channel support
Message-id: <1383636879-7067-2-git-send-email-jasowang@redhat.com>
Patchwork-id: 70565
O-Subject: [RHEL7.0 PATCH 01/21] Drivers: hv: vmbus: Implement multi-channel support
Bugzilla: 1026618
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>

From: K. Y. Srinivasan <kys@microsoft.com>

bugzilla: 1026618
test status: smoking test by myself
brew build: https://brewweb.devel.redhat.com/taskinfo?taskID=6525336

Starting with Win8, the host supports multiple sub-channels for a given
device. As in the past, the initial channel offer specifies the device and
is associated with both the type and the instance GUIDs. For performance
critical devices, the host may support multiple sub-channels. The sub-channels
share the same type and instance GUID as the primary channel. The number of
sub-channels offerrred to the guest depends on the number of virtual CPUs
assigned to the guest. The guest can request the creation of these sub-channels
and once created and opened, the guest can distribute the traffic across all
the channels (the primary and the sub-channels). A request sent on a sub-channel
will have the response delivered on the same sub-channel.

At channel (sub-channel) creation we bind the channel interrupt to a CPU and
with this sub-channel support we will be able to spread the interrupt load
of a given device across all available CPUs.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Reviewed-by: Haiyang Zhang <haiyangz@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit e68d2971d26577b932a16333ce165af98a96e068)

Signed-off-by: Jarod Wilson <jarod@redhat.com>