[redhat] Kconfig/s390x: switch off KVM for the kdump kernel variant
Message-id: <1383657528-7661-2-git-send-email-brueckner@redhat.com>
Patchwork-id: 70586
O-Subject: [RHEL7.0 PATCH] [redhat] s390/kconfig: switch off KVM for the kdump kernel variant
Bugzilla: 1026282
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
Turn off CONFIG_KVM for the kdump kernel variant as this is not required for
zfcpdump.  The zfcpdump kernel (which is a stand-alone dump utility; not for
kdump even if it is named that way) requires a small kernel image because of
low memory contraints at runtime.

Bugzilla
--------
BZ 1026282
https://bugzilla.redhat.com/show_bug.cgi?id=1026282

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=6526816

Signed-off-by: Jarod Wilson <jarod@redhat.com>