[scsi] scsi_dh_alua: ALUA handler attach should succeed while TPG is transitioning
Message-id: <1383772582-5885-3-git-send-email-emilne@redhat.com>
Patchwork-id: 70730
O-Subject: [RHEL7.0 PATCH e-stor 2/2] [SCSI] scsi_dh_alua: ALUA handler attach should succeed while TPG is transitioning
Bugzilla: 1020355
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

During testing, it was discovered that when a device tries to attach to the
alua handler while in TPG state of transitioning, the alua_rtpg function will
wait for it to exit the state before allowing it to continue. As a result, if
the 60 second timeout expires, the alua handler will not attach to the device.

To fix this, I have introduced an input argument to alua_rtpg called
wait_for_transition.  The idea is that it will wait for the transition to
complete before an activation (because the current TPG state has some bearing
in that case), but during a discovery if it is transitioning, it will not
wait, and will store the state as standby for the time being.

I believe the precedent exists for this from commit
c0d289b3e59577532c45ee9110ef81bd7b341272 Since if the device reports a state
of transitioning, it can transition to other more valid states, and it has
been established TPGS is supported on the device, if it is attaching.

Signed-off-by: Sean Stewart <Sean.Stewart@netapp.com>
Acked-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
(cherry picked from commit a8e5a2d593cbfccf530c3382c2c328d2edaa7b66)
Signed-off-by: Ewan D. Milne <emilne@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>