[fs] gfs2: Add allocation parameters structure
Message-id: <d7636d015ccc6ce53dc74403f68318644ecd9775.1384460582.git.rpeterso@redhat.com>
Patchwork-id: 70863
O-Subject: [RHEL7 PATCH 1/4] GFS2: Add allocation parameters structure
Bugzilla: 1019462
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

This patch adds a structure to contain allocation parameters with
the intention of future expansion of this structure. The idea is
that we should be able to add more information about the allocation
in the future in order to allow the allocator to make a better job
of placing the requests on-disk.

There is no functional difference from applying this patch.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>

rhbz#1019462

Signed-off-by: Jarod Wilson <jarod@redhat.com>