[fs] gfs2: If requested is too large, use the largest extent in the rgrp
Message-id: <b4b5fc3d0d818d9fb4d72fde13d3072dbabc62ef.1384460582.git.rpeterso@redhat.com>
Patchwork-id: 70864
O-Subject: [RHEL7 PATCH 2/4] GFS2: If requested is too large, use the largest extent in the rgrp
Bugzilla: 1019462
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Before this patch, GFS2 would keep searching through all the rgrps
until it found one that had a chunk of free blocks big enough to
satisfy the size hint, which is based on the file write size,
regardless of whether the chunk was big enough to perform the write.
However, when doing big writes there may not be a large enough
chunk of free blocks in any rgrp, due to file system fragmentation.
The largest chunk may be big enough to satisfy the write request,
but it may not meet the ideal reservation size from the "size hint".
The writes would slow to a crawl because every write would search
every rgrp, then finally give up and default to a single-block write.
In my case, performance would drop from 425MB/s to 18KB/s, or 24000
times slower.

This patch basically makes it so that if we can't find a contiguous
chunk of blocks big enough to satisfy the sizehint, we'll use the
largest chunk of blocks we found that will still contain the write.
It does so by keeping track of the largest run of blocks within the
rgrp.

rhbz#1019462

Signed-off-by: Jarod Wilson <jarod@redhat.com>