[fs] nfs: check if gssd is running before attempting to use krb5i auth in SETCLIENTID call
Message-id: <1384963006-15384-4-git-send-email-jlayton@redhat.com>
Patchwork-id: 71039
O-Subject: [RHEL7.0 PATCH 3/3] [BZ#998616] nfs: check if gssd is running before attempting to use krb5i auth in SETCLIENTID call
Bugzilla: 998616
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Currently, the client will attempt to use krb5i in the SETCLIENTID call
even if rpc.gssd isn't running. When that fails, it'll then fall back to
RPC_AUTH_UNIX. This introduced a delay when mounting if rpc.gssd isn't
running, and causes warning messages to pop up in the ring buffer.

Check to see if rpc.gssd is running before even attempting to use krb5i
auth, and just silently skip trying to do so if it isn't. In the event
that the admin is actually trying to mount with krb5*, it will still
fail at a later stage of the mount attempt.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream commit: 539d396704aa4780f84c88eaed9e1276bab12a23

Signed-off-by: Jarod Wilson <jarod@redhat.com>