[powerpc] kdump: Adding symbols in vmcoreinfo to facilitate dump filtering
Message-id: <20131127155645.44026.60497.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 71472
O-Subject: [PATCH RHEL7.0 BZ1025746] powerpc/kdump: Adding symbols in vmcoreinfo to facilitate dump filtering
Bugzilla: 1025746
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Chao WANG <chaowang@redhat.com>
RH-Acked-by: Dave Anderson <anderson@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1025746

Description:
------------
When CONFIG_SPARSEMEM_VMEMMAP option is used in kernel, makedumpfile fails
to filter vmcore dump as it fails to do vmemmap translations. So far
dump filtering on ppc64 never had to deal with vmemmap addresses seperately
as vmemmap regions where mapped in zone normal. But with the inclusion of
CONFIG_SPARSEMEM_VMEMMAP config option in kernel, this vmemmap address
translation support becomes necessary for dump filtering. For vmemmap adress
translation, few kernel symbols are needed by dump filtering tool. This patch
adds those symbols to vmcoreinfo, which a dump filtering tool can use for
filtering the kernel dump. Tested this changes successfully with makedumpfile
tool that supports vmemmap to physical address translation outside zone normal.

[ Removed unneeded #ifdef as suggested by Michael Ellerman --BenH ]

Signed-off-by: Hari Bathini <hbathini@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 7.0

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6653422

Upstream:
---------
http://git.kernel.org/cgit/linux/kernel/git/benh/powerpc.git/commit/?h=merge&id=8ff812719ac73ef9428b788e210221c8d87b2a8c

Test Status:
------------
I was able to successfully creat a vmcore. There are 2 changes for this to work, the first is this kernel change and
the second is makedumpfile.

On my system I have a test kexec-tools rpm with the the change for makedumpfile covered by bz 1031995. The change for
bz 1031995 is in the process of coming in.

kdump is broken on power at the moment so both the kernel and makedumpfile changes need to be committed
before it is functional again.

Paul Bunyan has tested these changes and they work for him also.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------
--
cgit v0.9.2

Signed-off-by: Jarod Wilson <jarod@redhat.com>