[scsi] bfa: Fix 1860 port initialize when ATC is enabled
Message-id: <1386745706-2047-10-git-send-email-vguvva@redhat.com>
Patchwork-id: 71846
O-Subject: [RHEL7.0 e-stor PATCH 09/24] scsi: bfa: Fix 1860 port initialize when ATC is enabled
Bugzilla: 726968
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>

Upstream Status:
e1aaab89dee184646f7001850e1fe6d55090a728

On Xen kernels, if ATC (address translation cache) is enabled, the
first PCIe DMA read from the adapter fails with an error. This is due to
a bug ASIC, which leads to a failure of 1860 ports to be initialised.

This patch includes the fix to disable Invalidated Tag Match Enable
capability by setting the bit 26 of CHIP_MISC_PRG to 0, by default it is
set to 1.

Signed-off-by: Anil Gurumurthy <agurumur@brocade.com>
Signed-off-by: Vijaya Mohan Guvva <vmohan@brocade.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>